src/crawler/requirements.txt
scrapy==2.11.0
playwright==1.40.0
scrapy-playwright==0.0.33
psycopg2-binary==2.9.9
python-dotenv==1.0.0
fake-useragent==1.4.0
